revert to total time filed integer

+++ office_tracker/daily_tracker/models.py
@@ -8,7 +8,7 @@ from .decorators import update_total_time
 class Attandance(TimeStampedModel):
     enter_at = models.DateTimeField(null=True)
     out_at = models.DateTimeField(null=True)
-    total_time = models.DecimalField(default=0, max_digits=19, decimal_places=10)
+    total_time = models.IntegerField(default=0)
     user = models.ForeignKey(User, on_delete=models.PROTECT, null=True)
 
     def is_logged_in(self):
